reshuffled the package structure for errors, added skeleton for error tests
