+++ src/main/scala/parsley/internal/errors/Builders.scala
@@ -15,18 +15,3 @@ private [internal] abstract class LineBuilder {
     protected def nearestNewlineAfter(off: Int): Int
     protected def segmentBetween(start: Int, end: Int): String
 }
-
-private [internal] abstract class ErrorItemBuilder {
-    final private [errors] def apply(offset: Int): ErrorItem = {
-        if (inRange(offset)) Raw(charAt(offset))
-        else EndOfInput
-    }
-    final private [errors] def apply(offset: Int, size: Int): ErrorItem = {
-        if (inRange(offset)) Raw(substring(offset, size))
-        else EndOfInput
-    }
-
-    protected def inRange(offset: Int): Boolean
-    protected def charAt(offset: Int): Char
-    protected def substring(offset: Int, size: Int): String
-}
