+++ src/main/scala/parsley/internal/machine/Context.scala
@@ -3,10 +3,9 @@ package parsley.internal.machine
 import instructions.Instr
 import stacks.{ArrayStack, Stack, CallStack, CheckStack, HandlerStack, StateStack, HintStack, ErrorStack}, Stack.StackExt
 import parsley.{Failure, Result, Success}
-import parsley.internal.errors.{
-    TrivialError,
-    ErrorItem, Desc,
-    LineBuilder, ErrorItemBuilder,
+import parsley.internal.errors.{ErrorItem, LineBuilder}
+import parsley.internal.machine.errors.{
+    ErrorItemBuilder,
     DefuncError, ClassicExpectedError, ClassicExpectedErrorWithReason, ClassicFancyError, ClassicUnexpectedError, WithHints,
     DefuncHints, EmptyHints, MergeHints, ReplaceHint, PopHints, AddError
 }
