+++ src/main/scala/parsley/internal/machine/instructions/ErrorInstrs.scala
@@ -3,7 +3,8 @@ package parsley.internal.machine.instructions
 import parsley.internal.machine.{Context, Good}
 import parsley.internal.ResizableArray
 
-import parsley.internal.errors.{ErrorItem, Desc, TrivialError, FailError, WithReason, MergedErrors, WithLabel}
+import parsley.internal.errors.{ErrorItem, Desc}
+import parsley.internal.machine.errors.{WithReason, MergedErrors, WithLabel}
 
 private [internal] final class ApplyError(label: String) extends Instr {
     val isHide: Boolean = label.isEmpty
