+++ src/main/scala/parsley/internal/machine/stacks/ErrorStack.scala
@@ -1,6 +1,6 @@
 package parsley.internal.machine.stacks
 
-import parsley.internal.errors.DefuncError
+import parsley.internal.machine.errors.DefuncError
 
 private [machine] final class ErrorStack(var error: DefuncError, val tail: ErrorStack)
 private [machine] object ErrorStack extends Stack[ErrorStack] {
