+++ src/test/scala/parsley/internal/machine/errors/DefuncErrorTests.scala
@@ -1,0 +1,11 @@
+package parsley.internal.machine.errors
+
+import parsley.ParsleyTest
+
+import scala.language.implicitConversions
+
+import MockedBuilders.mockedErrorItemBuilder
+
+class DefuncErrorTests extends ParsleyTest {
+    //TODO: Add tests to verify the defunc behaviours
+}
